# Conflicts:
#	src/acceleration/plugin_security.cpp
#	src/acceleration/vec_knn.cpp
#	src/config/config_schema_validator.cpp
#	src/core/concerns/redis_cache.cpp
#	src/plugins/plugin_manager.cpp
#	src/storage/backup_manager.cpp
#	src/storage/blob_redundancy_manager.cpp
#	src/storage/columnar_format.cpp
#	src/storage/compressed_storage.cpp
#	src/storage/streaming_ingest_manager.cpp
#	src/tensor/hyper_index_builder.cpp
#	src/tensor/tensor_fingerprint_graph.cpp
#	src/tensor/tensor_index_manager.cpp
#	src/tensor/utr_converter.cpp

+++ src/acceleration/nccl_vector_backend.cpp
@@ -577,8 +577,8 @@ bool NCCLVectorBackend::checkNVLinkSupport(const std::vector<int>& deviceIds) {
     }
     
     // Check if P2P is available between devices (simplified NVLink check)
-    for (size_t i = 0; i <static_cast<int>(deviceIds.size()); ++i) {
-        for (size_t j = i + 1; j <static_cast<int>(deviceIds.size()); ++j) {
+    for (size_t i = 0; i < deviceIds.size(); ++i) {
+        for (size_t j = i + 1; j < deviceIds.size(); ++j) {
             int canAccess = 0;
             cudaDeviceCanAccessPeer(&canAccess, deviceIds[i], deviceIds[j]);
             if (canAccess) {
