+++ src/aql/aql_rollback_suggester.cpp
@@ -80,7 +80,7 @@ std::string extractCollection(const std::string &upper, const std::string &in_ke
             // Skip space, read word.
             size_t ws = after + 1;
             size_t we = ws;
-            while ((we <static_cast<int>(upper.size()) && (std::isalnum(static_cast<unsigned char>(upper[we])) || upper[we] == '_'))) {
+            while ((we < upper.size() && (std::isalnum(static_cast<unsigned char>(upper[we])) || upper[we] == '_'))) {
                 ++we;
             }
             if (we > ws) {
@@ -103,7 +103,7 @@ std::string extractLoopVar(const std::string &upper) {
     }
     size_t vs = p + 4;
     size_t ve = vs;
-    while ((ve <static_cast<int>(upper.size()) && (std::isalnum(static_cast<unsigned char>(upper[ve])) || upper[ve] == '_'))) {
+    while ((ve < upper.size() && (std::isalnum(static_cast<unsigned char>(upper[ve])) || upper[ve] == '_'))) {
         ++ve;
     }
     if (ve > vs) {
