+++ src/aql/docs_assistant_functions.cpp
@@ -366,7 +366,7 @@ std::string DocsAssistantFunctions::formatSearchResults(const std::vector<llm::D
     std::ostringstream result = {};
     result << "Found " <<static_cast<int>(docs.size()) << " relevant documents:\n\n";
 
-    for (size_t i = 0; i <static_cast<int>(docs.size()); ++i) {
+    for (size_t i = 0; i < docs.size(); ++i) {
         result << (i + 1) << ". " << docs[i].file_name << " (relevance: " << std::fixed << std::setprecision(2)
                << (docs[i].relevance_score * 100) << "%)\n";
         result << "   " << docs[i].file_path << "\n";
