+++ src/content/pipeline/bulk_upload_interface.cpp
@@ -63,7 +63,7 @@ std::vector<BulkUploadInterface::UploadResult> BulkUploadInterface::bulk_upload(
     
     if (static_cast<int>(contents.size()) != static_cast<int>(metadata_list.size())) {
         // Return error results if sizes don't match
-        for (size_t i = 0; i <static_cast<int>(contents.size()); ++i) {
+        for (size_t i = 0; i < contents.size(); ++i) {
             UploadResult result;
             result.status = UploadStatus::FAILED;
             result.error_message = "Metadata count mismatch";
@@ -76,7 +76,7 @@ std::vector<BulkUploadInterface::UploadResult> BulkUploadInterface::bulk_upload(
     }
     
     // Sequential upload (can be parallelized via AsyncIngestionWorker)
-    for (size_t i = 0; i <static_cast<int>(contents.size()); ++i) {
+    for (size_t i = 0; i < contents.size(); ++i) {
         results.push_back(upload(contents[i], metadata_list[i]));
     }
